refactoring effort for clarity and topical separation
